feat(types): add analytics event definition
