+++ packages/types/src/AnalyticsEvent.ts
@@ -1,0 +1,18 @@
+export type AnalyticsEvent =
+  | {
+      type: "email_sent" | "email_open" | "email_click";
+      campaign: string;
+      timestamp?: string;
+      [key: string]: unknown;
+    }
+  | {
+      type: "discount_redeemed";
+      code: string;
+      timestamp?: string;
+      [key: string]: unknown;
+    }
+  | {
+      type: string;
+      timestamp?: string;
+      [key: string]: unknown;
+    };
